+++ NEWS.md
@@ -2,4 +2,5 @@
 # abscorr 0.1.0
 
 * use pkgdown for documentation
-* Added a `NEWS.md` file to track changes to the package.
+* Added a `NEWS.md` file to track changes to the package
+* Import SEIFA scores for various spatial geometries